+++ test/clj_http/test/client.clj
@@ -1,6 +1,6 @@
 (ns clj-http.test.client
   (:require [clj-http.lite.client :as client]
-            [clj-http.test.core :refer [base-req with-server]]
+            [clj-http.test.core :refer [base-req with-server current-port]]
             [clj-http.lite.util :as util]
             [clojure.test :refer [deftest is testing use-fixtures]])
   (:import (java.net UnknownHostException)))
@@ -18,6 +18,41 @@
                                                 :scheme "http"}))]
     (is (= 200 (:status resp)))
     (is (= "get" (:body resp)))))
+
+(deftest ^{:integration true} basic-auth-no-creds
+  (let [resp (client/request (merge (base-req) {:method :get
+                                                :uri "/basic-auth"
+                                                :throw-exceptions false}))]
+    (is (= 401 (:status resp)))
+    (is (= "denied" (:body resp)))))
+
+(deftest ^{:integration true} basic-auth-bad-creds
+  (let [resp (client/request (merge (base-req) {:method :get
+                                                :uri "/basic-auth"
+                                                :throw-exceptions false
+                                                :basic-auth "username:nope"}))]
+    (is (= 401 (:status resp)))
+    (is (= "denied" (:body resp)))))
+
+(deftest ^{:integration true} basic-auth-creds-as-basic-auth
+  (let [resp (client/request (merge (base-req) {:method :get
+                                                :uri "/basic-auth"
+                                                :basic-auth "username:password"}))]
+    (is (= 200 (:status resp)))
+    (is (= "welcome" (:body resp)))))
+
+(deftest ^{:integration true} basic-auth-creds-as-user-info
+  (let [resp (client/request (merge (base-req) {:method :get
+                                                :uri "/basic-auth"
+                                                :user-info "username:password"}))]
+    (is (= 200 (:status resp)))
+    (is (= "welcome" (:body resp)))))
+
+(deftest ^{:integration true} basic-auth-creds-from-url
+  (let [resp (client/request {:method :get
+                              :url (format "http://username:password@localhost:%d/basic-auth" (current-port))})]
+    (is (= 200 (:status resp)))
+    (is (= "welcome" (:body resp)))))
 
 (defn is-passed [middleware req]
   (let [client (middleware identity)]